+++ apps/avatax/src/modules/webhooks/calculate-taxes-payload.ts
@@ -1,0 +1,9 @@
+import { CalculateTaxesEventFragment } from "../../../generated/graphql";
+
+/**
+ * Common subset for ORDER and CHECKOUT events
+ */
+export type CalculateTaxesPayload = Extract<
+  CalculateTaxesEventFragment,
+  { __typename: "CalculateTaxes" }
+>;