+++ apps/avatax/src/pages/api/webhooks/checkout-calculate-taxes.ts
@@ -1,37 +1,23 @@
 import { SaleorSyncWebhook } from "@saleor/app-sdk/handlers/next";
-import {
-  CalculateTaxesEventFragment,
-  UntypedCalculateTaxesDocument,
-} from "../../../../generated/graphql";
+import { UntypedCalculateTaxesDocument } from "../../../../generated/graphql";
 import { saleorApp } from "../../../../saleor-app";
 import { WebhookResponse } from "../../../modules/app/webhook-response";
 import { getActiveConnectionService } from "../../../modules/taxes/get-active-connection-service";
-import { TaxIncompleteWebhookPayloadError } from "../../../modules/taxes/tax-error";
 import { withOtel } from "@saleor/apps-otel";
 import { createLogger } from "../../../logger";
+import * as Sentry from "@sentry/nextjs";
+import { calculateTaxesErrorsStrategy } from "../../../modules/webhooks/calculate-taxes-errors-strategy";
+
+import { verifyCalculateTaxesPayload } from "../../../modules/webhooks/validate-webhook-payload";
+import { CalculateTaxesPayload } from "../../../modules/webhooks/calculate-taxes-payload";
+import { wrapWithLoggerContext } from "@saleor/apps-logger/node";
+import { loggerContext } from "../../../logger-context";
 
 export const config = {
   api: {
     bodyParser: false,
   },
 };
-
-export type CalculateTaxesPayload = Extract<
-  CalculateTaxesEventFragment,
-  { __typename: "CalculateTaxes" }
->;
-
-function verifyCalculateTaxesPayload(payload: CalculateTaxesPayload) {
-  if (!payload.taxBase.lines.length) {
-    throw new TaxIncompleteWebhookPayloadError("No lines found in taxBase");
-  }
-
-  if (!payload.taxBase.address) {
-    throw new TaxIncompleteWebhookPayloadError("No address found in taxBase");
-  }
-
-  return payload;
-}
 
 export const checkoutCalculateTaxesSyncWebhook = new SaleorSyncWebhook<CalculateTaxesPayload>({
   name: "CheckoutCalculateTaxes",
@@ -41,34 +27,72 @@ export const checkoutCalculateTaxesSyncWebhook = new SaleorSyncWebhook<Calculate
   webhookPath: "/api/webhooks/checkout-calculate-taxes",
 });
 
-export default withOtel(
-  checkoutCalculateTaxesSyncWebhook.createHandler(async (req, res, ctx) => {
-    const logger = createLogger("checkoutCalculateTaxesSyncWebhook");
-    const { payload } = ctx;
-    const webhookResponse = new WebhookResponse(res);
+/**
+ * TODO: Add tests to handler
+ */
+export default wrapWithLoggerContext(
+  withOtel(
+    checkoutCalculateTaxesSyncWebhook.createHandler(async (req, res, ctx) => {
+      const logger = createLogger("checkoutCalculateTaxesSyncWebhook");
+      const { payload } = ctx;
+      const webhookResponse = new WebhookResponse(res);
 
-    logger.info("Handler for CHECKOUT_CALCULATE_TAXES webhook called");
+      logger.info("Handler for CHECKOUT_CALCULATE_TAXES webhook called");
 
-    try {
-      verifyCalculateTaxesPayload(payload);
-      logger.debug("Payload validated Successfully");
+      try {
+        const payloadVerificationResult = verifyCalculateTaxesPayload(payload);
 
-      const appMetadata = payload.recipient?.privateMetadata ?? [];
-      const channelSlug = payload.taxBase.channel.slug;
-      const activeConnectionService = getActiveConnectionService(
-        channelSlug,
-        appMetadata,
-        ctx.authData,
-      );
+        if (payloadVerificationResult.isErr()) {
+          logger.debug("Failed to calculate taxes, due to incomplete payload", {
+            error: payloadVerificationResult.error,
+          });
 
-      logger.info("Found active connection service. Calculating taxes...");
-      const calculatedTaxes = await activeConnectionService.calculateTaxes(payload);
+          return res.status(400).send(payloadVerificationResult.error.message);
+        }
 
-      logger.info("Taxes calculated", { calculatedTaxes });
-      return webhookResponse.success(ctx.buildResponse(calculatedTaxes));
-    } catch (error) {
-      return webhookResponse.error(error);
-    }
-  }),
-  "/api/webhooks/checkout-calculate-taxes",
+        const appMetadata = payload.recipient?.privateMetadata ?? [];
+        const channelSlug = payload.taxBase.channel.slug;
+        const activeConnectionServiceResult = getActiveConnectionService(
+          channelSlug,
+          appMetadata,
+          ctx.authData,
+        );
+
+        if (activeConnectionServiceResult.isErr()) {
+          const err = activeConnectionServiceResult.error;
+
+          logger.debug(`Error in taxes calculation occurred: ${err.name} ${err.message}`, {
+            error: err,
+          });
+
+          const executeErrorStrategy = calculateTaxesErrorsStrategy(req, res).get(err.name);
+
+          if (executeErrorStrategy) {
+            return executeErrorStrategy();
+          } else {
+            Sentry.captureException(err);
+
+            logger.fatal(`UNHANDLED: ${err.name}`, {
+              error: err,
+            });
+
+            return res.status(500).send("Error calculating taxes");
+          }
+        } else {
+          logger.debug("Found active connection service. Calculating taxes...");
+          // TODO: Improve errors handling like above
+          const calculatedTaxes = await activeConnectionServiceResult.value.calculateTaxes(payload);
+
+          logger.debug("Taxes calculated", { calculatedTaxes });
+          return webhookResponse.success(ctx.buildResponse(calculatedTaxes));
+        }
+      } catch (error) {
+        Sentry.captureException(error);
+
+        return webhookResponse.error(error);
+      }
+    }),
+    "/api/webhooks/checkout-calculate-taxes",
+  ),
+  loggerContext,
 );