+++ apps/avatax/src/pages/api/webhooks/order-cancelled.ts
@@ -8,6 +8,8 @@ import { getActiveConnectionService } from "../../../modules/taxes/get-active-co
 import { WebhookResponse } from "../../../modules/app/webhook-response";
 import { withOtel } from "@saleor/apps-otel";
 import { createLogger } from "../../../logger";
+import { wrapWithLoggerContext } from "@saleor/apps-logger/node";
+import { loggerContext } from "../../../logger-context";
 export const config = {
   api: {
     bodyParser: false,
@@ -27,35 +29,49 @@ export const orderCancelledAsyncWebhook = new SaleorAsyncWebhook<OrderCancelledP
   webhookPath: "/api/webhooks/order-cancelled",
 });
 
-export default withOtel(
-  orderCancelledAsyncWebhook.createHandler(async (req, res, ctx) => {
-    const logger = createLogger("orderCancelledAsyncWebhook", {
-      saleorApiUrl: ctx.authData.saleorApiUrl,
-    });
-    const { payload } = ctx;
-    const webhookResponse = new WebhookResponse(res);
+export default wrapWithLoggerContext(
+  withOtel(
+    orderCancelledAsyncWebhook.createHandler(async (req, res, ctx) => {
+      const logger = createLogger("orderCancelledAsyncWebhook", {
+        saleorApiUrl: ctx.authData.saleorApiUrl,
+      });
+      const { payload } = ctx;
+      const webhookResponse = new WebhookResponse(res);
 
-    logger.info("Handler called with payload");
+      logger.info("Handler called with payload");
 
-    if (!payload.order) {
-      return webhookResponse.error(new Error("Insufficient order data"));
-    }
+      if (!payload.order) {
+        return webhookResponse.error(new Error("Insufficient order data"));
+      }
 
-    try {
-      const appMetadata = payload.recipient?.privateMetadata ?? [];
-      const channelSlug = payload.order.channel.slug;
-      const taxProvider = getActiveConnectionService(channelSlug, appMetadata, ctx.authData);
+      try {
+        const appMetadata = payload.recipient?.privateMetadata ?? [];
+        const channelSlug = payload.order.channel.slug;
+        const taxProviderResult = getActiveConnectionService(
+          channelSlug,
+          appMetadata,
+          ctx.authData,
+        );
 
-      logger.info("Cancelling order...");
+        logger.info("Cancelling order...");
 
-      await taxProvider.cancelOrder(payload);
+        if (taxProviderResult.isOk()) {
+          await taxProviderResult.value.cancelOrder(payload);
 
-      logger.info("Order cancelled");
+          logger.info("Order cancelled");
 
-      return webhookResponse.success();
-    } catch (error) {
-      return webhookResponse.error(error);
-    }
-  }),
-  "/api/webhooks/order-cancelled",
+          return webhookResponse.success();
+        }
+
+        if (taxProviderResult.isErr()) {
+          // TODO: Map errors
+          return webhookResponse.error(taxProviderResult.error);
+        }
+      } catch (error) {
+        return webhookResponse.error(error);
+      }
+    }),
+    "/api/webhooks/order-cancelled",
+  ),
+  loggerContext,
 );