+++ packages/shared/src/create-graphql-client.ts
@@ -1,9 +1,12 @@
 import { authExchange } from "@urql/exchange-auth";
-import { cacheExchange, createClient as urqlCreateClient, fetchExchange } from "urql";
+import { cacheExchange, createClient as urqlCreateClient, fetchExchange, Exchange } from "urql";
 
-interface CreateGraphQLClientArgs {
+export interface CreateGraphQLClientArgs {
   saleorApiUrl: string;
   token?: string;
+  opts?: {
+    prependingFetchExchanges?: Exchange[];
+  };
 }
 
 /*
@@ -17,7 +20,13 @@ interface CreateGraphQLClientArgs {
  *
  * In the context of developing Apps, the two first options are recommended.
  */
-export const createGraphQLClient = ({ saleorApiUrl, token }: CreateGraphQLClientArgs) => {
+export const createGraphQLClient = ({ saleorApiUrl, token, opts }: CreateGraphQLClientArgs) => {
+  const beforeFetch = [];
+
+  if (opts?.prependingFetchExchanges) {
+    beforeFetch.push(...opts?.prependingFetchExchanges);
+  }
+
   return urqlCreateClient({
     url: saleorApiUrl,
     exchanges: [
@@ -39,6 +48,7 @@ export const createGraphQLClient = ({ saleorApiUrl, token }: CreateGraphQLClient
           async refreshAuth() {},
         };
       }),
+      ...beforeFetch,
       fetchExchange,
     ],
   });