+++ apps/crm/CHANGELOG.md
@@ -1,5 +1,12 @@
 # saleor-app-crm
 
+## 1.8.4
+
+### Patch Changes
+
+- Updated dependencies [67afe8e4]
+  - @saleor/apps-shared@1.9.3
+
 ## 1.8.3
 
 ### Patch Changes