+++ apps/emails-and-messages/CHANGELOG.md
@@ -1,5 +1,22 @@
 # saleor-app-emails-and-messages
 
+## 1.11.2
+
+### Patch Changes
+
+- Updated dependencies [1e07a6ff]
+  - @saleor/apps-logger@1.2.0
+
+## 1.11.1
+
+### Patch Changes
+
+- 67afe8e4: Apps that use OTEL can now collect and send spans containing details about GraphQL requests.
+- Updated dependencies [67afe8e4]
+- Updated dependencies [67afe8e4]
+  - @saleor/apps-shared@1.9.3
+  - @saleor/apps-otel@1.1.0
+
 ## 1.11.0
 
 ### Minor Changes